requirements.txt
accelerate==0.18.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
brotlipy==0.7.0
certifi==2022.12.7
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
debugpy @ file:///C:/ci_310/debugpy_1642079916595/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
diffusers==0.15.1
docopt==0.6.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fastcore==1.5.29
filelock @ file:///C:/b/abs_c7yrhs9uz2/croot/filelock_1672387617533/work
huggingface-hub==0.13.4
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1682176699712/work
ipykernel @ file:///D:/bld/ipykernel_1655369313836/work
ipython @ file:///D:/bld/ipython_1680185618122/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1681432441054/work
jupyter_core @ file:///D:/bld/jupyter_core_1678994400417/work
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
mpmath==1.2.1
mypy-extensions==1.0.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx @ file:///C:/b/abs_b935xy_9g6/croot/networkx_1678964342510/work
numpy @ file:///C:/b/abs_datssh7cer/croot/numpy_and_numpy_base_1672336199388/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.4.0
pipreqs==0.4.13
platformdirs @ file:///C:/b/abs_73cc5cz_1u/croots/recipe/platformdirs_1662711386458/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
psutil==5.9.5
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyre-extensions==0.0.23
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pywin32==305.1
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1657616000714/work
regex==2023.3.23
requests @ file:///C:/b/abs_80_edxtvhu/croot/requests_1678709752318/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.0
tornado @ file:///D:/bld/tornado_1656937966227/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformers==4.28.1
typing-inspect==0.8.0
typing_extensions @ file:///C:/b/abs_a1bb332wcs/croot/typing_extensions_1681939523095/work
urllib3 @ file:///C:/b/abs_3ce53vrdcr/croot/urllib3_1680254693505/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
xformers==0.0.18
yarg==0.1.9
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
